Refactored and packaged code for exercise
